+++ examples/spatial_3d.rs
@@ -105,8 +105,8 @@ fn change_velocity(
 ) {
     let mut emitter = query.single_mut();
 
-    let x = time.elapsed_seconds_wrapped().sin() as f32 * 3.0;
-    let z = time.elapsed_seconds_wrapped().cos() as f32 * 3.0;
+    let x = time.elapsed_seconds_wrapped().sin() * 3.0;
+    let z = time.elapsed_seconds_wrapped().cos() * 3.0;
     let delta = time.delta_seconds();
 
     let sink = match sinks.get_mut(&sink.0) {
